Update get http and ws endpoint methods. Move methods common to multiple API types to a common file
