+++ adapter/internal/oasparser/model/api_operation.go
@@ -37,34 +37,34 @@ type Operation struct {
 	disableSecurity  bool
 	vendorExtensions map[string]interface{}
 	policies         OperationPolicies
-	mockedAPIConfig  MockedAPIConfig    
+	mockedAPIConfig  MockedAPIConfig
 }
 
 // MockedAPIConfig holds configurations defined for a mocked API operation result
 type MockedAPIConfig struct {
-	In          string                    `json:"in,omitempty"`
-	Name        string                    `json:"name,omitempty"`
-	Responses   []MockedResponseConfig    `json:"responses,omitempty"`
+	In        string                 `json:"in,omitempty"`
+	Name      string                 `json:"name,omitempty"`
+	Responses []MockedResponseConfig `json:"responses,omitempty"`
 }
 
 // MockedResponseConfig holds response configurations in the mocked API operation
 type MockedResponseConfig struct {
-	Value 	string                        `json:"value,omitempty"`
-	Headers []MockedHeaderConfig          `json:"headers,omitempty"`
-	Code	int                           `json:"code,omitempty"`	
-	Content []MockedContentConfig         `json:"content,omitempty"`
+	Value   string                `json:"value,omitempty"`
+	Headers []MockedHeaderConfig  `json:"headers,omitempty"`
+	Code    int                   `json:"code,omitempty"`
+	Content []MockedContentConfig `json:"content,omitempty"`
 }
 
 // MockedHeaderConfig holds header configurations in the mocked API operation
 type MockedHeaderConfig struct {
-	Name  string                          `json:"name,omitempty"`
-	Value string                          `json:"value,omitempty"`
+	Name  string `json:"name,omitempty"`
+	Value string `json:"value,omitempty"`
 }
 
 // MockedContentConfig holds mocked content configurations in the mocked API operation
 type MockedContentConfig struct {
-	ContentType string                    `json:"contentType,omitempty"`
-	Body string                           `json:"body,omitempty"`
+	ContentType string `json:"contentType,omitempty"`
+	Body        string `json:"body,omitempty"`
 }
 
 // GetMethod returns the http method name of the give API operation
@@ -131,7 +131,7 @@ func (operation *Operation) GetCallInterceptorService(isIn bool) InterceptEndpoi
 					if urlFound {
 						url, isString := urlValue.(string)
 						if isString && url != "" {
-							endpoint, err := getHostandBasepathandPort(url)
+							endpoint, err := getHTTPEndpoint(url)
 							if err == nil {
 								conf, _ := config.ReadConfigs()
 								clusterTimeoutV := conf.Envoy.ClusterTimeoutInSeconds
@@ -164,7 +164,7 @@ func (operation *Operation) GetCallInterceptorService(isIn bool) InterceptEndpoi
 }
 
 // NewOperation Creates and returns operation type object
-func NewOperation(method string, security []map[string][]string, extensions map[string]interface{}, 
+func NewOperation(method string, security []map[string][]string, extensions map[string]interface{},
 	mockedAPIConfig MockedAPIConfig) *Operation {
 	tier := ResolveThrottlingTier(extensions)
 	disableSecurity := ResolveDisableSecurity(extensions)
