+++ adapter/internal/oasparser/model/swagger_internal_test.go
@@ -128,7 +128,7 @@ func TestSetResourcesSwagger(t *testing.T) {
 		},
 	}
 	for _, item := range dataItems {
-		resultResources := setResourcesSwagger(item.input, false)
+		resultResources := getResourcesSwagger(item.input, false)
 		if item.result != nil {
 			assert.Equal(t, item.result[0].path, resultResources[0].GetPath(), item.message)
 			resultResources[0].GetMethod()[0].iD = item.result[0].methods[0].iD
